+++ src/components/news/news-message/index.tsx
@@ -1,0 +1,2 @@
+export * from './news-message';
+export { default } from './news-message';