+++ src/components/news/news-page/index.tsx
@@ -1,0 +1,2 @@
+export * from './news-page';
+export { default } from './news-page';